# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MoneyGuard API
  description: |
    # MoneyGuard API
    
    API для приложения управления личными финансами MoneyGuard, которое позволяет пользователям отслеживать доходы и расходы, 
    устанавливать бюджеты, создавать финансовые цели и получать аналитику по своим финансам.
    
    ## Аутентификация
    
    Все защищенные эндпоинты требуют JWT токен, который можно получить через `/auth/login` или `/auth/register`.
    Токен необходимо передавать в заголовке `Authorization` в формате `Bearer {token}`.
    
    ## Доступ к ИИ
    
    Приложение предоставляет базовый функционал для всех пользователей. Доступ к AI-рекомендациям можно получить,
    активировав соответствующую опцию через API подписки.
  version: 1.0.0
  contact:
    email: 123
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен авторизации, полученный при регистрации или входе в систему
  parameters:
...
    limitId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Уникальный идентификатор лимита
      example: 1
...
  schemas:
...
    SubscriptionResponse:
      type: object
      properties:
        ai_access_enabled:
          type: boolean
          description: Флаг доступа к AI-рекомендациям
          example: true
        features:
          type: array
          items:
            type: string
          description: Список доступных функций
          example:
            - Базовый учет доходов и расходов
            - Неограниченное количество категорий
            - Неограниченное количество транзакций
            - Доступ к AI-рекомендациям
...
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор категории
          example: 1
        name:
          type: string
          description: Название категории
          example: Продукты
        parent_id:
          type: integer
          nullable: true
          description: ID родительской категории (null для корневых категорий)
          example: null
        icon:
          type: string
          description: Иконка категории
          example: shopping_cart
...
    LimitRequest:
      type: object
      required: [ category_id, amount, period ]
      properties:
        category_id:
          type: integer
          description: ID категории, для которой устанавливается лимит
          example: 1
        amount:
          type: number
          minimum: 0
          description: Сумма лимита (должна быть положительной)
          example: 5000
        period:
          type: string
          enum: [ DAILY, WEEKLY, MONTHLY, YEARLY ]
          description: Период действия лимита
          example: MONTHLY
    LimitResponse:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор лимита
          example: 1
        category:
          $ref: '#/components/schemas/CategoryResponse'
        amount:
          type: number
          description: Сумма лимита
          example: 5000
        period:
          type: string
          description: Период действия лимита
          example: MONTHLY
        current_spending:
          type: number
          description: Текущие расходы по категории за указанный период
          example: 2500
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP статус код ошибки
          example: 404
        message:
          type: string
          description: Сообщение об ошибке
          example: Категория с ID 999 не найдена
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: 2023-04-01T12:00:00Z
        path:
          type: string
          description: Путь запроса, вызвавшего ошибку
          example: /api/categories/999
    ValidationErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP статус код ошибки
          example: 400
        message:
          type: string
          description: Общее сообщение об ошибке
          example: Ошибка валидации данных
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: 2023-04-01T12:00:00Z
        path:
          type: string
          description: Путь запроса, вызвавшего ошибку
          example: /api/categories
        errors:
          type: object
          additionalProperties:
            type: string
          description: Детальные ошибки валидации по полям
          example:
            name: Название категории не может быть пустым
            icon: Иконка должна быть указана
paths:
...
  /subscription:
    get:
      tags: [ Subscription ]
      summary: Получить информацию о подписке
      description: |
        Возвращает информацию о текущей подписке пользователя, включая статус доступа к AI-рекомендациям.
        
        Требуется JWT токен для авторизации.
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Информация о подписке
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
              examples:
                example1:
                  summary: Пример ответа с информацией о подписке без доступа к ИИ
                  value:
                    ai_access_enabled: false
                    features:
                      - Базовый учет доходов и расходов
                      - Неограниченное количество категорий
                      - Неограниченное количество транзакций
                example2:
                  summary: Пример ответа с информацией о подписке с доступом к ИИ
                  value:
                    ai_access_enabled: true
                    features:
                      - Базовый учет доходов и расходов
                      - Неограниченное количество категорий
                      - Неограниченное количество транзакций
                      - Доступ к AI-рекомендациям
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [ Subscription ]
      summary: Включить доступ к AI-рекомендациям
      description: |
        Включает доступ к AI-рекомендациям для текущего пользователя.
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Доступ к AI-рекомендациям успешно включен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
              examples:
                example1:
                  summary: Пример ответа после включения доступа к ИИ
                  value:
                    ai_access_enabled: true
                    features:
                      - Базовый учет доходов и расходов
                      - Неограниченное количество категорий
                      - Неограниченное количество транзакций
                      - Доступ к AI-рекомендациям
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [ Subscription ]
      summary: Отключить доступ к AI-рекомендациям
      description: |
        Отключает доступ к AI-рекомендациям для текущего пользователя.
      security:
        - bearerAuth: [ ]
      responses:
        '204':
          description: Доступ к AI-рекомендациям успешно отключен
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /limits:
    get:
      tags: [ Limits ]
      summary: Получить список лимитов
      description: |
        Возвращает список лимитов расходов текущего пользователя.
        
        Лимиты позволяют контролировать расходы по категориям за определенный период.
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Список лимитов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LimitResponse'
              examples:
                example1:
                  summary: Пример ответа со списком лимитов
                  value:
                    - id: 1
                      category:
                        id: 1
                        name: Продукты
                        parent_id: null
                        icon: shopping_cart
                      amount: 10000
                      period: MONTHLY
                      current_spending: 5000
                    - id: 2
                      category:
                        id: 3
                        name: Развлечения
                        parent_id: null
                        icon: movie
                      amount: 5000
                      period: MONTHLY
                      current_spending: 2000
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [ Limits ]
      summary: Создать новый лимит
      description: |
        Создает новый лимит расходов для указанной категории.
        
        Для каждой категории может быть установлен только один лимит на определенный период.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LimitRequest'
            examples:
              example1:
                summary: Пример запроса на создание лимита
                value:
                  category_id: 1
                  amount: 10000
                  period: MONTHLY
      responses:
        '201':
          description: Лимит успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitResponse'
              examples:
                example1:
                  summary: Пример ответа с созданным лимитом
                  value:
                    id: 1
                    category:
                      id: 1
                      name: Продукты
                      parent_id: null
                      icon: shopping_cart
                    amount: 10000
                    period: MONTHLY
                    current_spending: 0
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Категория не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Лимит для этой категории уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                example1:
                  summary: Пример ошибки при создании дублирующего лимита
                  value:
                    status: 409
                    message: Лимит для этой категории уже существует
                    timestamp: 2023-04-01T12:00:00Z
                    path: /limits
  /limits/{id}:
    get:
      tags: [ Limits ]
      summary: Получить лимит по ID
      description: |
        Возвращает информацию о лимите по его ID.
        
        Лимит должен принадлежать текущему пользователю.
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/limitId'
      responses:
        '200':
          description: Информация о лимите
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitResponse'
              examples:
                example1:
                  summary: Пример ответа с информацией о лимите
                  value:
                    id: 1
                    category:
                      id: 1
                      name: Продукты
                      parent_id: null
                      icon: shopping_cart
                    amount: 10000
                    period: MONTHLY
                    current_spending: 5000
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Лимит не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [ Limits ]
      summary: Обновить лимит
      description: |
        Обновляет информацию о лимите.
        
        Лимит должен принадлежать текущему пользователю.
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/limitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LimitRequest'
            examples:
              example1:
                summary: Пример запроса на обновление лимита
                value:
                  category_id: 1
                  amount: 15000
                  period: MONTHLY
      responses:
        '200':
          description: Лимит успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitResponse'
              examples:
                example1:
                  summary: Пример ответа с обновленным лимитом
                  value:
                    id: 1
                    category:
                      id: 1
                      name: Продукты
                      parent_id: null
                      icon: shopping_cart
                    amount: 15000
                    period: MONTHLY
                    current_spending: 5000
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Лимит или категория не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [ Limits ]
      summary: Удалить лимит
      description: |
        Удаляет лимит по его ID.
        
        Лимит должен принадлежать текущему пользователю.
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/limitId'
      responses:
        '204':
          description: Лимит успешно удален
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Лимит не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...